add shared mocks
