+++ test/helpers/helpers.dart
@@ -7,4 +7,5 @@
 
 export 'hydrated_bloc.dart';
 export 'ignore_errors.dart';
+export 'mocks.dart';
 export 'pump_app.dart';
